backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
libclang==16.0.6
sqlalchemy==2.0.25
pydantic==2.5.3
networkx==3.2.1
python-multipart==0.0.6
